Add files via upload
The purpose of this was to test out the given code and to evaluate if the return() function worked.  Going of the code above, we can see that it does indeed work due to the fact that the function of return() is to recall the last function, which it successfully does.